refactor(formatter): split formatter facts domains
